fix(ci): split audit inspection handlers
